Continued refactoring free question visualizations
