+++ notebooks/configuration.json
@@ -6,6 +6,8 @@
     "figures_save_dir_autogen": "./test/figs/auto_gen"
   },
   "visualizations": {
-    "single": [{ "finishers": true, "chart": "bar", "ordered": true, "likert": false }]
+    "single": [{ "finishers": true, "chart": "bar", "ordered": true, "likert": true }],
+    "free": [{ "plot": "wordcloud" }],
+    "freematrix": [{ "plot": "textlist" }]
   }
 }
